move calculation of route probabilities to static class Utiltiy.Random
