+++ Assets/Scripts/DataTypes/Vertices/EndPoint.cs
@@ -11,9 +11,8 @@ namespace DataTypes
         private GameObject _carPrefab { get; }
         // ticks before a car spawns on a lane (index)
         private Frequencies _frequencies { get; }
-        // possible routing destinations with weight
-        private RouteProbabilities _routeProbabilities;
-        private int[] _weights { get; }
+        // cumulative Probabilities of choosing a vertex to route to
+        private List<double> _cumulativeProbabilities { get; }
         public Dictionary<IVertex, List<RouteSegment>> routingTable { get; } = new Dictionary<IVertex, List<RouteSegment>>();
         private static readonly HashSet<LaneType> _ONLY_THROUGH = new HashSet<LaneType> {LaneType.Through};
 
@@ -22,22 +21,19 @@ namespace DataTypes
             _edge = edge;
             _carPrefab = carPrefab;
             _frequencies = frequencies;
-            _weights = weights;
+            _cumulativeProbabilities = MathUtils.CalculateCumulative(weights);
 
             if (edge.incomingLanes.Any(lane => !lane.types.Equals(_ONLY_THROUGH)))
                 throw new NetworkConfigurationError("All lanes going into an EndPoint have to be of type Through");
-        }
-
-        public void CalculateRouteProbabilities()
-        {
-            _routeProbabilities = new RouteProbabilities(_weights, routingTable.Where(kvp => kvp.Value != null).Select(kvp => kvp.Key).ToList());
         }
 
         public void SpawnCars()
         {
             foreach (var lane in _frequencies.CurrentActiveIndices())
             {
-                new Car(_carPrefab, lane, routingTable[_routeProbabilities.Choose()].ToList());
+                new Car(_carPrefab, lane, routingTable[Utility.Random.Choose(
+                    cumulativeProbabilities: _cumulativeProbabilities, 
+                    destinations: routingTable.Where(kvp => kvp.Value != null).Select(kvp => kvp.Key).ToList())].ToList());
             }
         }
